+++ changelog.md
@@ -1,9 +1,25 @@
 # Version log
 
 ## 0.0.1
+### issue#6 (Tue 19 Apr 15:42:37 BST 2022)
+- beam selection from plotly
+- add rpx,rpy,Ipets data to the plotly graph
+- show miller indices when hovering over data points
+- add a callback which adds/removes in a javascript array the selected hkl
+- use and show proper markers
+- add a background color (light blue) to the graph
+- add px and py labels to the axes
+- persistent xylims
+- persistent data visibility
+- adding MathJax dependency
+- added structure info in `crys`
+- bloch solver layout with expandable parameters
+- bug fix when updating frame orientation vector  
+- added rotation parameter omega
+- added selected beams to rocking curve and thickness dep
 
 ### issue#8 (Mon 11 Apr 16:53:08 BST 2022)
-- Rocking curve 
+- Rocking curve
 - [x] compute and display rocking curve in side figure
 - [x] add panel for rocking curve parameters
 - [x] show computation evolution